+++ Models/Usta_Table.cs
@@ -27,7 +27,7 @@ namespace UstaYardımAPI.Models
         
         [ForeignKey("MahalleinfoMahalleId")]
         public Mahalleler? Mahalleinfo { get; set; }  // Sehir tablosuna referans
-        public int? Puan { get; set; }
+        public double? Puan { get; set; }
         public string? Hakkinda { get; set; }
         public DateTime? Birthday { get; set; }
         public string? TamamlananIs { get; set; }
